Generalized the call for user added sheets.
Running the command from cheat_sheet from ST will now take an argument with the name of the desired sheet. CheatSheedCommand will check ./Packages/User and then the plugin directory.

I removed the read only and name change because it was an unnecessary nuisnace - a cheatsheet should be improved, and a well named cheat sheet's purpose will be obvious in the tab (ie Regular Expression.cheatsheet.)

I also added a cheat_sheet_tester function that will print to ST's console what I looked for and whether or not it found it.